Feat/responsive menu (#58)
* chore: fix logo props type

* chore: update MUI dependencies

* chore: type use user preferences hook

* feat: add collapsed menu prototype

* feat: add collapsed menu icon and styling

* chore: fix build with ts-ignores

* fix: disable scroll lock to prevent scrollbar jumping

* fix: link should fill menu item
