+++ package.json
@@ -21,10 +21,10 @@
     "@fingerprintjs/fingerprintjs-pro": "^3.6.2",
     "@fingerprintjs/fingerprintjs-pro-react": "^2.3.0",
     "@fingerprintjs/fingerprintjs-pro-server-api": "^2.1.0",
-    "@mui/icons-material": "^5.8.4",
-    "@mui/lab": "^5.0.0-alpha.89",
-    "@mui/material": "^5.8.7",
-    "@mui/styles": "^5.8.7",
+    "@mui/icons-material": "^5.11.16",
+    "@mui/lab": "^5.0.0-alpha.134",
+    "@mui/material": "^5.13.5",
+    "@mui/styles": "^5.13.2",
     "clsx": "^1.2.1",
     "leaflet": "^1.9.4",
     "next": "^13.1.6",
